+++ src/app/component/conversation-detail/conversation-header/conversation-header.component.ts
@@ -23,6 +23,7 @@ export class ConversationHeaderComponent implements OnInit, OnChanges {
   @Input() typingLocation: string;
   @Input() isTrascriptDownloadEnabled: boolean;
   @Input() fullscreenMode: boolean;
+  @Input() size: 'min' | 'max' | 'top';
   @Input() hideCloseConversationOptionMenu: boolean;
   @Input() hideHeaderCloseButton: boolean;
   @Input() hideHeaderBackButton: boolean;
@@ -104,10 +105,9 @@ export class ConversationHeaderComponent implements OnInit, OnChanges {
     this.onMenuOptionClick.emit(HEADER_MENU_OPTION.LOGOUT)
   }
   /**
-   * @param status : string 'max' || 'min'
+   * @param status : string 'max' || 'min' || 'top'
    */
   onChangeSize(status){
-    this.heightStatus = status
     if(status === 'min') this.onMenuOptionClick.emit(HEADER_MENU_OPTION.MINIMIZE)
     if(status === 'max') this.onMenuOptionClick.emit(HEADER_MENU_OPTION.MAXIMIZE)
     if(status === 'top') this.onMenuOptionClick.emit(HEADER_MENU_OPTION.TOP)