+++ src/app/providers/global-settings.service.ts
@@ -367,15 +367,19 @@ export class GlobalSettingsService {
         }
         // console.log('this.globals.fullscreenMode' + this.globals.fullscreenMode);
         if (this.globals.fullscreenMode === true) {
-            // divTiledeskiframe.style.left = 0;
-            // divTiledeskiframe.style.right = 0;
-            // divTiledeskiframe.style.top = 0;
-            // divTiledeskiframe.style.bottom = 0;
-            // divTiledeskiframe.style.width = '100%';
-            // divTiledeskiframe.style.height = '100%';
-            // divTiledeskiframe.style.maxHeight = 'none';
-            // divTiledeskiframe.style.maxWidth = 'none';
-            divTiledeskiframe.classList.add('fullscreen')
+            divTiledeskiframe.style.left = 0;
+            divTiledeskiframe.style.right = 0;
+            divTiledeskiframe.style.top = 0;
+            divTiledeskiframe.style.bottom = 0;
+            divTiledeskiframe.style.width = '100%';
+            divTiledeskiframe.style.height = '100%';
+            divTiledeskiframe.style.maxHeight = 'none';
+            divTiledeskiframe.style.maxWidth = 'none';
+            // divTiledeskiframe.classList.add('fullscreen')
+            divTiledeskiframe.classList.remove('min-size')
+            divTiledeskiframe.classList.remove('max-size')
+            divTiledeskiframe.classList.remove('top-size')
+
         }
     }
 
@@ -547,6 +551,9 @@ export class GlobalSettingsService {
                     }
                     if (variables.hasOwnProperty('allowedOnSpecificUrlList')) {
                         globals['allowedOnSpecificUrlList'] = variables['allowedOnSpecificUrlList'];
+                    }
+                    if (variables.hasOwnProperty('allowedUploadExtentions')) {
+                        globals['fileUploadAccept'] = variables['allowedUploadExtentions'];
                     }
                     
                 }
@@ -915,7 +922,7 @@ export class GlobalSettingsService {
             globals.hideSettings = (TEMP === true) ? true : false;;
         }
         TEMP = tiledeskSettings['isLogEnabled'];
-        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > logLevel:: ', TEMP]);
+        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > isLogEnabled:: ', TEMP]);
         if (TEMP !== undefined) {
             globals.isLogEnabled = TEMP;
         }
@@ -964,7 +971,7 @@ export class GlobalSettingsService {
             globals.fontFamily = TEMP + ',' + globals.fontFamily;
         }
         TEMP = tiledeskSettings['fontFamilySource'];
-        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > fontFamily:: ', TEMP]);
+        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > fontFamilySource:: ', TEMP]);
         if (TEMP !== undefined) {
             globals.fontFamilySource = TEMP;
         }
@@ -1013,7 +1020,7 @@ export class GlobalSettingsService {
             globals.nativeRating = (TEMP === true) ? true : false;
         }
         TEMP = tiledeskSettings['showInfoMessage'];
-        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > showBubbleInfoMessage:: ', TEMP]);
+        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > showInfoMessage:: ', TEMP]);
         if (TEMP !== undefined) {
             globals.showInfoMessage = TEMP.split(',').map(key => { return key.trim()});
         }
@@ -1056,7 +1063,7 @@ export class GlobalSettingsService {
             globals.telegramUsername = TEMP;
         }
         TEMP = tiledeskSettings['fileUploadAccept'];
-        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > telegramUsername:: ', TEMP]);
+        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > fileUploadAccept:: ', TEMP]);
         if (TEMP !== undefined) {
             globals.fileUploadAccept = TEMP;
         } 
@@ -1096,9 +1103,14 @@ export class GlobalSettingsService {
             globals.showEmojiFooterButton = (TEMP === true) ? true : false;
         } 
         TEMP = tiledeskSettings['showAudioRecorderFooterButton'];
-        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > showEmojiFooterButton:: ', TEMP]);
+        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > showAudioRecorderFooterButton:: ', TEMP]);
         if (TEMP !== undefined) {
             globals.showAudioRecorderFooterButton = (TEMP === true) ? true : false;
+        }
+        TEMP = tiledeskSettings['size'];
+        // this.logger.debug('[GLOBAL-SET] setVariablesFromSettings > size:: ', TEMP]);
+        if (TEMP !== undefined) {
+            globals.size = TEMP;
         } 
     }
 
@@ -1840,6 +1852,11 @@ export class GlobalSettingsService {
         TEMP = getParameterByName(windowContext, 'tiledesk_showEmojiFooterButton');
         if (TEMP) {
             globals.showEmojiFooterButton = stringToBoolean(TEMP);
+        }
+
+        TEMP = getParameterByName(windowContext, 'tiledesk_size');
+        if (TEMP) {
+            globals.size = TEMP;
         }
         
     }